Removed unnecessary retrofit instances
